removed legacy file

+++ components/exchange/buy/limit.tsx
@@ -95,7 +95,7 @@ const Limit = ({
                         ...buySellLimitCoinData,
                         price: e.target.value,
                         total:
-                          parseInt(e.target.value) *
+                          parseFloat(e.target.value) *
                           buySellLimitCoinData.amount,
                       });
                     }}
@@ -122,7 +122,8 @@ const Limit = ({
                         ...buySellLimitCoinData,
                         amount: e.target.value,
                         total:
-                          parseInt(e.target.value) * buySellLimitCoinData.price,
+                          parseFloat(e.target.value) *
+                          buySellLimitCoinData.price,
                       });
                     }}
                   />
