+++ components/exchange/sell/stopLimit.tsx
@@ -143,7 +143,7 @@ const StopLimit = ({
                         ...buySellStopLimitCoinData,
                         amount: e.target.value,
                         total:
-                          parseInt(e.target.value) *
+                          parseFloat(e.target.value) *
                           dashboard?.order_data?.sell_price,
                       });
                       // await dispatch(
